+++ index.mjs
@@ -1,27 +1,21 @@
 import { writable, derived, get as readStore } from "svelte/store";
 export default function(origins, derive, reflect, initial) {
-	var originValues, allowReflect = true, allowDerive = true;
-	var childWritable = writable(initial);
-	var childWritableSetter = childWritable.set;
-	var sendDownstream = (value) => {
-		allowReflect = false;
-		childWritableSetter(value);
-		allowReflect = true;
-	};
+	var childDerivedSetter, originValues, allowDerive = true;
 	var wrappedDerive = (got, set) => {
+		childDerivedSetter = set;
 		originValues = got;
 		if (allowDerive) {
-			let returned = derive(got, sendDownstream);
+			let returned = derive(got, set);
 			if (derive.length < 2) {
-				sendDownstream(returned);
+				set(returned);
 			} else {
 				return returned;
 			}
 		}
 	};
-	var childDerived = derived(origins, wrappedDerive);
+	var childDerived = derived(origins, wrappedDerive, initial);
 	
-	var singleOrigin = !Array.isArray(origins), unsubscribeFromDerived;
+	var singleOrigin = !Array.isArray(origins);
 	var sendUpstream = (setWith) => {
 		allowDerive = false;
 		if (singleOrigin) {
@@ -34,72 +28,51 @@ export default function(origins, derive, reflect, initial) {
 		allowDerive = true;
 	};
 	var cleanup = null;
-	allowReflect = false;
-	childWritable.subscribe((value) => {
-		if (allowReflect) {
-			if (cleanup) {
-				cleanup();
-				cleanup = null;
-			}
-			
-			let isAsync = false;
-			let returned = reflect({
-				reflecting: value,
-				get old() {
-					// We need an active subscription to childDerived for originValues to be usable
-					if (unsubscribeFromDerived) {
-						return originValues;
-					} else {
-						if (singleOrigin) {
-							return readStore(origins);
-						} else {
-							return origins.map(readStore);
-						}
-					}
-				},
-				get set() {
-					isAsync = true;
-					return sendUpstream;
-				},
-			});
-			if (isAsync) {
-				if (typeof returned == "function") {
-					cleanup = returned;
-				}
-			} else {
-				sendUpstream(returned);
-			}
+	function doReflect(value) {
+		if (cleanup) {
+			cleanup();
+			cleanup = null;
 		}
-	});
-	allowReflect = true;
-	function listen() {
-		unsubscribeFromDerived = childDerived.subscribe( () => {} );
-	}
-	function unlisten() {
-		unsubscribeFromDerived();
-		originValues = unsubscribeFromDerived = undefined;
+
+		let isAsync = false;
+		let returned = reflect({
+			reflecting: value,
+			old: originValues,
+			get set() {
+				isAsync = true;
+				return sendUpstream;
+			},
+		});
+		if (isAsync) {
+			if (typeof returned == "function") {
+				cleanup = returned;
+			}
+		} else {
+			sendUpstream(returned);
+		}
 	}
 	
-	var subscriberCount = 0;
-	var me = {
-		subscribe(subscriber) {
-			++subscriberCount;
-			if (subscriberCount == 1) { listen(); }
-			var unsubscribe = childWritable.subscribe(subscriber);
-			return () => {
-				unsubscribe();
-				--subscriberCount;
-				if (subscriberCount == 0) { unlisten(); }
-			};
-		},
-		set: childWritable.set,
-		update(fn) {
-			if (subscriberCount == 0) {
-				// guarantee up-to-date value
-				listen(), unlisten();
+	var activeUpdateId = 0;
+	function update(fn) {
+		var firstCall = true, isUpdated, updateId = ++activeUpdateId, oldValue;
+		var unsubscribe = childDerived.subscribe( (value) => {
+			if (firstCall) {
+				oldValue = value;
+				firstCall = false;
+			} else {
+				isUpdated = true;
 			}
-			childWritable.update(fn);
-		},
+		} );
+		var newValue = fn(oldValue);
+		childDerivedSetter(newValue);
+		unsubscribe();
+		if (isUpdated && updateId == activeUpdateId) {
+			doReflect(newValue);
+		}
+	}
+	return {
+		subscribe: childDerived.subscribe,
+		set(value) { update( () => value ); },
+		update,
 	};
-	return me;
 }